Update FRU bus data type to uint16_t
The FruDevice service changed the type of FRU's bus index from uint8_t
to uint16_t which consequently made the "bus" parameter of "GetRawFru"
and "WriteFru" methods updated to uint16_t.
This commit updates the type of FRU's bus to uint16_t to adapt with
FruDevice service.

Tested:
1. Request to update MAC address
$ipmitool raw 0x3c 0x01 0xa1 0xa2 0xa3 0xa4 0xa5 0xa6
3. Reboot then check new MAC address
4. New MAC address is updated

+++ src/oemcommands.cpp
@@ -87,7 +87,7 @@ static inline auto responseSetFanErrorThermalCtlNotDisabled()
  *  @param - busIdx, address of I2C device
  *  @returns - true if successfully, false if fail
  */
-[[maybe_unused]] static bool getBaseBoardFRUAddr(uint8_t &busIdx, uint8_t &addr)
+[[maybe_unused]] static bool getBaseBoardFRUAddr(uint16_t &busIdx, uint8_t &addr)
 {
     bool retVal = false;
     sd_bus* bus = NULL;
@@ -146,7 +146,7 @@ static inline auto responseSetFanErrorThermalCtlNotDisabled()
                     (busProperty != fruDeviceInf->second.end()) &&
                     (addrProperty != fruDeviceInf->second.end()))
                 {
-                    busIdx = (uint8_t)std::get<uint32_t>(busProperty->second);
+                    busIdx = (uint16_t)std::get<uint32_t>(busProperty->second);
                     addr = (uint8_t)std::get<uint32_t>(addrProperty->second);
                     retVal = true;
                     break;
@@ -164,7 +164,7 @@ static inline auto responseSetFanErrorThermalCtlNotDisabled()
  *         - fruData: data have been read
  *  @returns - true if successfully, false if fail
  */
-static bool getRawFruData(uint8_t busIdx, uint8_t addr, std::vector<uint8_t> &fruData)
+static bool getRawFruData(uint16_t busIdx, uint8_t addr, std::vector<uint8_t> &fruData)
 {
     bool retVal = false;
     sd_bus* bus = NULL;
@@ -287,7 +287,7 @@ static bool updateMACAddInFRU(std::vector<uint8_t> &fruData, std::vector<uint8_t
  *         - fruData: FRU data
  *  @returns - true if successfully, false if fail
  */
-static bool writeFruData(uint8_t busIdx, uint8_t addr, std::vector<uint8_t> &fruData)
+static bool writeFruData(uint16_t busIdx, uint8_t addr, std::vector<uint8_t> &fruData)
 {
     bool retVal = false;
     sd_bus* bus = NULL;
@@ -543,7 +543,7 @@ ipmi::RspType<> ipmiDocmdScpWriteRegisterMap(uint8_t cpuIndex, uint8_t offsetW,
 ipmi::RspType<uint8_t> ipmiDocmdSetMacAddress(std::vector<uint8_t> macAddress)
 {
     std::vector<uint8_t> fruData;
-    uint8_t busIdx = 0;
+    uint16_t busIdx = 0;
     uint8_t addrss = 0;
 
     if(macAddress.size() != 6)
